+++ xcube/core/optimize.py
@@ -21,7 +21,7 @@
 
 import os.path
 import shutil
-from typing import Type
+from typing import Type, Union, Sequence
 
 import zarr
 
@@ -31,7 +31,7 @@ from xcube.core.unchunk import unchunk_dataset
 def optimize_dataset(input_path: str,
                      output_path: str = None,
                      in_place: bool = False,
-                     unchunk_coords: bool = False,
+                     unchunk_coords: Union[bool, str, Sequence[str]] = False,
                      exception_type: Type[Exception] = ValueError):
     """
     Optimize a dataset for faster access.
@@ -40,17 +40,21 @@ def optimize_dataset(input_path: str,
     Consolidated cubes open much faster from remote locations, e.g. in object storage,
     because obviously much less HTTP requests are required to fetch initial cube meta
     information. That is, it merges all metadata files into a single top-level JSON file ".zmetadata".
-    If *unchunk_coords* is set, it also removes any chunking of coordinate variables
+
+    If *unchunk_coords* is given, it also removes any chunking of coordinate variables
     so they comprise a single binary data file instead of one file per data chunk.
     The primary usage of this function is to optimize data cubes for cloud object storage.
-    The function currently works only for data cubes using ZARR format.
+    The function currently works only for data cubes using Zarr format.
+    *unchunk_coords* can be a name, or list of names of the coordinate variable(s) to be consolidated.
+    If boolean ``True`` is used, coordinate all variables will be consolidated.
 
     :param input_path: Path to input dataset with ZARR format.
     :param output_path: Path to output dataset with ZARR format. May contain "{input}" template string,
-           which is replaced by the input path's file name without file name extentsion.
+           which is replaced by the input path's file name without file name extension.
     :param in_place: Whether to modify the dataset in place.
            If False, a copy is made and *output_path* must be given.
-    :param unchunk_coords: Whether to also consolidate coordinate chunk files.
+    :param unchunk_coords: The name of a coordinate variable or a list of coordinate variables whose chunks should
+           be consolidated. Pass ``True`` to consolidate chunks of all coordinate variables.
     :param exception_type: Type of exception to be used on value errors.
     """
 
@@ -75,6 +79,12 @@ def optimize_dataset(input_path: str,
         shutil.copytree(input_path, output_path)
 
     if unchunk_coords:
-        unchunk_dataset(output_path, coords_only=True)
+        if isinstance(unchunk_coords, str):
+            var_names = (unchunk_coords,)
+        elif isinstance(unchunk_coords, bool):
+            var_names = None
+        else:
+            var_names = tuple(unchunk_coords)
+        unchunk_dataset(output_path, var_names=var_names, coords_only=True)
 
     zarr.convenience.consolidate_metadata(output_path)