+++ CHANGES.md
@@ -110,6 +110,10 @@
 
 ### Other
 
+* The `unchunk_coords` keyword argument of Python API function 
+  `xcube.core.optimize.optimize_dataset()` can now be a name, or list of names  
+  of the coordinate variable(s) to be consolidated. If boolean ``True`` is used
+  all variables will be consolidated.
 * The `xcube serve` API operations `datasets/` and `datasets/{ds_id}` now also
   return the metadata attributes of a given dataset and it variables in a property
   named `attrs`. For variables we added a new metadata property `htmlRepr` that is